MyJavaHomework1

+++ src/Main.java
@@ -7,27 +7,27 @@ public class Main {
 
         System.out.println("Hello Raul!");
 
-        int adunare = 7 + 8;
+        byte adunare = 7 + 8;
         System.out.println(adunare);
         System.out.println(" Rezultatul adunarii dintre 7 si 8 este: " + adunare);
 
-        int impartire = 26 / 2;
+        byte impartire = 26 / 2;
         System.out.println(impartire);
         System.out.println(" Rezultatul impartirii dintre 26 si 2 este: " + impartire);
 
-        int ecuatie1 = (-5 + 8 * 6);
+        byte ecuatie1 = (-5 + 8 * 6);
         System.out.println(ecuatie1 );
         System.out.println(" Rezultatul ecuatie 1 este: " + ecuatie1);
 
-        int ecuatie2 = ((55 + 9) % 9);
+        byte ecuatie2 = ((55 + 9) % 9);
         System.out.println(ecuatie2);
         System.out.println(" Rezultatul ecuatie 2 este: " + ecuatie2);
 
-        int ecuatie3 = (20 + (((-3) * 5) / 8 ));
+        float ecuatie3 = (20 + (((-3) * 5) / 8 ));
         System.out.println(ecuatie3);
         System.out.println(" Rezultatul ecuatie 3 este: " + ecuatie3);
 
-        int ecuatie4 = ( 5 + 15 / 3 * 2 - 8 % 3);
+        byte ecuatie4 = ( 5 + 15 / 3 * 2 - 8 % 3);
         System.out.println(ecuatie4);
         System.out.println(" Rezultatul ecuatie 4 este: " + ecuatie4);
